requirements_updated.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.39
botocore==1.42.39
branca==0.8.2
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
curl_cffi==0.13.0
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
filetype==1.2.0
finnhub-python==2.4.27
folium==0.20.0
fredapi==0.5.2
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.67.0
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.12
langchain-anthropic==1.3.4
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.16.0
numpy==2.4.2
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.1
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.5
psycopg2==2.9.11
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.50.0
stockstats==0.6.8
streamlit==1.54.0
streamlit-folium==0.26.1
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
watchdog==6.0.0
websockets==16.0
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.23.0
yfinance==1.1.0
zstandard==0.25.0
